root_orchestrator/resource-abstractor/requirements.txt
flask~=2.1.3
flask_pymongo~=2.3.0
flask-smorest~=0.39.0
flask-swagger-ui~=4.11.1
marshmallow~=3.15.0
werkzeug==2.0.3
paho-mqtt==1.6.1